Add files via upload
when running, the valid inputs go in this order:
The first line selects a search structure. It can be any of the following:
EMPTY, which specifies an "empty set", a search structure that stores no keys and for which all searches return false. (This is used primarily in performance testing.)
LIST, which specifies an unsorted linked list implementation (which has been provided already).
HASH ZERO, which specifies a hash table implementation using a hash function that always returns zero.
HASH SUM, which specifies a hash table implementation using a hash function that sums the character codes for every character of a word.
HASH PRODUCT, which specifies a hash table implementation using a hash function that includes multiplication by prime numbers in its calculation.
AVL, which specifies an AVL tree.
The second line specifies the path to a file that contains the word set (i.e., the set of correctly-specified words). The word set file will be assumed to have one correctly-spelled word on each line.
possibly a large set of English words obtained from an open source wordlist project called wordlist.sourceforge.net.
You might also want to create one or more smaller word sets to test with, since the provided word set has over 60,000 words.
The third line specifies the path to a file that contains the input text for which you want to check spelling. 
The fourth line specifies the format of the output you'd like to see, which allows you to choose between seeing a display of the misspellings and suggestions, or whether you're just interested in seeing how much time the spell-checking process takes. This line needs to say either:
DISPLAY, if you want a display of misspellings and suggestions.
TIME, if you want to see only timing results.